Move sidebar settings option below home button option
