+++ browser/resources/settings/brave_appearance_page/BUILD.gn
@@ -16,6 +16,7 @@ js_type_check("closure_compile_module") {
     ":brave_theme",
     ":super_referral",
     ":toolbar",
+    ":sidebar",
   ]
 }
 
@@ -74,6 +75,15 @@ js_library("toolbar") {
   ]
 }
 
+js_library("sidebar") {
+  externs_list = [
+    "$externs_path/settings_private.js",
+  ]
+  extra_deps = [
+    ":templatize"
+  ]
+}
+
 # make modules
 group("web_modules") {
   public_deps = [
@@ -84,6 +94,7 @@ group("web_modules") {
 html_to_js("templatize") {
   js_files = [
     "brave_theme.js",
+    "sidebar.js",
     "super_referral.js",
     "toolbar.js",
   ]
