+++ browser/resources/settings/brave_appearance_page/sidebar.html
@@ -1,0 +1,19 @@
+<style include="settings-shared iron-flex">
+</style>
+<template is="dom-if" if="[[isSidebarFeatureEnabled_]]">
+  <div class="settings-box">
+    <div class="flex" id="labelWrapper" aria-hidden="true">
+      <div class="label">$i18n{appearanceSettingsShowOptionTitle}</div>
+      <div class="secondary label" id="sub-label">
+        <span id="sub-label-text" aria-hidden="true">
+          [[sidebarShowEnabledLabel_]]
+        </span>
+      </div>
+    </div>
+    <settings-dropdown-menu
+      pref="{{prefs.brave.sidebar.sidebar_show_option}}"
+      on-settings-control-change="onShowOptionChanged_"
+      menu-options="[[sidebarShowOptions_]]">
+    </settings-toggle-button>
+  </div>
+</template>
