+++ browser/resources/settings/brave_appearance_page/toolbar.html
@@ -1,28 +1,10 @@
-<style include="settings-shared iron-flex">
+<style include="settings-shared">
 </style>
 <settings-toggle-button
   class="cr-row"
   pref="{{prefs.brave.location_bar_is_wide}}"
   label="$i18n{appearanceSettingsLocationBarIsWide}">
 </settings-toggle-button>
-
-<template is="dom-if" if="[[isSidebarFeatureEnabled_]]">
-  <div class="settings-box">
-    <div class="flex" id="labelWrapper" aria-hidden="true">
-      <div class="label">$i18n{appearanceSettingsShowOptionTitle}</div>
-      <div class="secondary label" id="sub-label">
-        <span id="sub-label-text" aria-hidden="true">
-          [[sidebarShowEnabledLabel_]]
-        </span>
-      </div>
-    </div>
-    <settings-dropdown-menu
-      pref="{{prefs.brave.sidebar.sidebar_show_option}}"
-      on-settings-control-change="onShowOptionChanged_"
-      menu-options="[[sidebarShowOptions_]]">
-    </settings-toggle-button>
-  </div>
-</template>
 <settings-toggle-button
   class="cr-row"
   pref="{{prefs.brave.autocomplete_enabled}}"
